--- deployment/docker-compose.yml
@@ -1,67 +1,0 @@
-version: "3.7"
-services:
-    # store
-    store:
-        build: ../cluster-companion/db
-        networks: 
-            - big-store-net
-
-    # companion 1
-    cluster-companion-1:
-        build: ../cluster-companion
-        restart: on-failure
-        environment:
-            - CACHE_HOST=cache-1
-        ports:
-            - 60011:60001
-        networks:
-            - big-store-net
-            - cluster-companion-1-net
-    cache-1:
-        image: redis
-        # ports:
-        #     - 6379:6379
-        networks:
-            - cluster-companion-1-net
-
-    # companion 2
-    cluster-companion-2:
-        build: ../cluster-companion
-        restart: on-failure
-        environment:
-            - CACHE_HOST=cache-2
-        ports:
-            - 60012:60001
-        networks:
-            - big-store-net
-            - cluster-companion-2-net
-    cache-2:
-        image: redis
-        # ports:
-        #     - 6379:6379
-        networks:
-            - cluster-companion-2-net
-
-    # companion 3
-    cluster-companion-3:
-        build: ../cluster-companion
-        restart: on-failure
-        environment:
-            - CACHE_HOST=cache-3
-        ports:
-            - 60013:60001
-        networks:
-            - big-store-net
-            - cluster-companion-3-net
-    cache-3:
-        image: redis
-        # ports:
-        #     - 6379:6379
-        networks:
-            - cluster-companion-3-net
-
-networks:
-    big-store-net:
-    cluster-companion-1-net:
-    cluster-companion-2-net:
-    cluster-companion-3-net:
